fix warnings and bugs fd_geometry

+++ src/simu/boundary.h
@@ -3,6 +3,6 @@
 struct boundary_point_s {
 	long index[N_boundary_point_s];
 	_Complex float val;
-	int dir[N_boundary_point_s];
+	long dir[N_boundary_point_s];
 };
 #endif
